+++ DefaultTabBar.js
@@ -1,17 +1,16 @@
-const React = require('react');
-const { ViewPropTypes } = ReactNative = require('react-native');
-const PropTypes = require('prop-types');
-const createReactClass = require('create-react-class');
-const {
+import React, {Component} from 'react';
+import {
   StyleSheet,
   Text,
   View,
   Animated,
-} = ReactNative;
+  ViewPropTypes,
+} from 'react-native';
+const PropTypes = require('prop-types');
 const Button = require('./Button');
 
-const DefaultTabBar = createReactClass({
-  propTypes: {
+export default class DefaultTabBar extends Component {
+  static propTypes = {
     goToPage: PropTypes.func,
     activeTab: PropTypes.number,
     tabs: PropTypes.array,
@@ -24,23 +23,22 @@ const DefaultTabBar = createReactClass({
     tabStyle: ViewPropTypes.style,
     renderTab: PropTypes.func,
     underlineStyle: ViewPropTypes.style,
-  },
+  };
 
-  getDefaultProps() {
-    return {
-      activeTextColor: 'navy',
-      inactiveTextColor: 'black',
-      activeTextFontSize: 25,
-      inactiveTextFontSize: 25,
-      backgroundColor: null,
-    };
-  },
+  static defaultProps = {
+    activeTextColor: 'navy',
+    inactiveTextColor: 'black',
+    activeTextFontSize: 25,
+    inactiveTextFontSize: 25,
+    backgroundColor: null,
+  };
 
-  renderTabOption(name, page) {
-  },
+  renderTabOption = (name, page) => {
 
-  renderTab(name, page, isTabActive, onPressHandler) {
-    const { activeTextColor, inactiveTextColor, textStyle,activeTextFontSize,inactiveTextFontSize } = this.props;
+  };
+
+  renderTab = (name, page, isTabActive, onPressHandler) => {
+    const { activeTextColor, inactiveTextColor, textStyle, activeTextFontSize, inactiveTextFontSize} = this.props;
     const textColor = isTabActive ? activeTextColor : inactiveTextColor;
     const textFontSize = isTabActive? activeTextFontSize : inactiveTextFontSize;
     const fontWeight = isTabActive ? 'bold' : 'normal';
@@ -59,7 +57,7 @@ const DefaultTabBar = createReactClass({
         </Text>
       </View>
     </Button>;
-  },
+  };
 
   render() {
     const containerWidth = this.props.containerWidth;
@@ -67,7 +65,6 @@ const DefaultTabBar = createReactClass({
     const tabUnderlineStyle = {
       position: 'absolute',
       width: containerWidth / numberOfTabs,
-      height: 4,
       backgroundColor: 'navy',
       bottom: 0,
     };
@@ -77,27 +74,23 @@ const DefaultTabBar = createReactClass({
       outputRange: [0,  containerWidth / numberOfTabs],
     });
     return (
-      <View style={[styles.tabs, {backgroundColor: this.props.backgroundColor, }, this.props.style, ]}>
+      <View style={[styles.tabs, {backgroundColor: this.props.backgroundColor, }, this.props.style ]}>
         {this.props.tabs.map((name, page) => {
           const isTabActive = this.props.activeTab === page;
           const renderTab = this.props.renderTab || this.renderTab;
           return renderTab(name, page, isTabActive, this.props.goToPage);
         })}
-        <Animated.View
-          style={[
-            tabUnderlineStyle,
-            {
-              transform: [
-                { translateX },
-              ]
-            },
-            this.props.underlineStyle,
-          ]}
-        />
+        <Animated.View style={[tabUnderlineStyle,{backgroundColor:'transparent'}, { transform: [
+            { translateX },
+          ] }]} >
+          <View style={[tabUnderlineStyle, {height: 4}, this.props.underlineStyle ]} >
+
+          </View>
+        </Animated.View>
       </View>
     );
-  },
-});
+  };
+}
 
 const styles = StyleSheet.create({
   tab: {
@@ -117,5 +110,3 @@ const styles = StyleSheet.create({
     borderColor: '#ccc',
   },
 });
-
-module.exports = DefaultTabBar;