+++ CHANGELOG.md
@@ -1,8 +1,21 @@
+# 1.0.6 (2021-03-17)
+
+* 修复对`@react-native-community/viewpager`高版本的兼容
+
+# 1.0.5 (2020-08-04)
+
+* scrollableTabBar的underline改为非撑满模式
+
+# 1.0.4
+
+* 修复RN0.62版本下，报getNode的警告问题
+
 # 1.0.3
-* 修复DefaultTabBar报错
 
+* 修复RN0.62版本下，报getNode的警告问题
+* 修复ScrollableTabBar的`activeTextFontSize`与`inactiveTextFontSize`
 
-# 1.0.2
+# 1.0.1
 
 * 添加activeTextFontSize和inactiveTextFontSize两个属性
 * 改造ScrollableTabBar的underline结构，保持跟DefaultTabBar一致